Reorganise cewe2pdf.py into smaller units, phase 5 Collect text area processing to separate file
